feat(dvc): initialize DVC with GCS remote (US-002)

+++ src/acquisition/tile_stitcher.py
@@ -169,4 +169,4 @@ class TileStitcher:
         a = np.sin(delta_phi / 2) ** 2 + np.cos(phi1) * np.cos(phi2) * np.sin(delta_lambda / 2) ** 2
         c = 2 * np.arctan2(np.sqrt(a), np.sqrt(1 - a))
 
-        return earth_radius * c
+        return float(earth_radius * c)
